rename hash file
