+++ app/utils/hash.js
